Fixed issue with 'cd' explanation. Added 'pwd'
The instructions for `cd` seemed to really imply `pwd`, so I changed that definition to that.   `cd` without an argument changes to the home directory, which is probably confusing.  

I switched the order of `cd ..` and `cd directory`, as its easier to explain how to switch *into* a directory, and then how to switch out of it.   If this makes sense, go for it.